+++ modules/backend/behaviors/FormController.php
@@ -108,6 +108,17 @@ class FormController extends ControllerBehavior
         if ($this->controller->isClassExtendedWith('Backend.Behaviors.RelationController')) {
             $this->controller->initRelation($model);
         }
+
+        $this->prepareVars($model);
+    }
+
+    /**
+     * Prepares common form data
+     */
+    protected function prepareVars($model)
+    {
+        $this->controller->vars['formModel'] = $model;
+        $this->controller->vars['formRecordName'] = Lang::get($this->getConfig('name', 'backend::lang.model.name'));
     }
 
     //
@@ -129,7 +140,6 @@ class FormController extends ControllerBehavior
             );
             $model = $this->controller->formCreateModelObject();
             $this->initForm($model);
-
             $this->controller->vars['formModel'] = $model;
         } catch (Exception $ex) {
             $this->controller->handleError($ex);
@@ -184,7 +194,7 @@ class FormController extends ControllerBehavior
             );
             $model = $this->controller->formFindModelObject($recordId);
             $this->initForm($model);
-
+            
             $this->controller->vars['formModel'] = $model;
         } catch (Exception $ex) {
             $this->controller->handleError($ex);
@@ -262,7 +272,7 @@ class FormController extends ControllerBehavior
             );
             $model = $this->controller->formFindModelObject($recordId);
             $this->initForm($model);
-
+            
             $this->controller->vars['formModel'] = $model;
         } catch (Exception $ex) {
             $this->controller->handleError($ex);
@@ -368,10 +378,10 @@ class FormController extends ControllerBehavior
     {
         $name = $this->getConfig($name, $default);
         $vars = [
-            'name' => Lang::get($this->getConfig('name', trans('backend::lang.model.name')))
+            'name' => Lang::get($this->getConfig('name', 'backend::lang.model.name'))
         ];
         $vars = array_merge($vars, $extras);
-        return trans($name, $vars);
+        return Lang::get($name, $vars);
     }
 
     //